refactor(publisher): keep bestEffortNotify out of the public barrel (r3 3877646541)
The containment helper moves to an internal non-barrel module; the built
package root no longer exposes it (verified against dist). Behavior proof
unchanged: the throwing/rejecting-listener rows still pass.
