+++ packages/publisher/src/publisher.ts
@@ -217,23 +217,6 @@ export type ParticipantSignatureProvider = (
 ) => Promise<ReceiverSignature[]>;
 
 
-/**
- * r2 (3877540214) — THE one implementation of non-fail-closed listener dispatch: neither a
- * synchronous throw nor an asynchronous rejection may escape into the caller. Every
- * `onPublishConfirmed` invocation goes through here so the containment recipe cannot drift
- * between call sites.
- */
-export function bestEffortNotify<T>(
-  listener: ((arg: T) => void | Promise<void>) | undefined,
-  arg: T,
-): void {
-  try {
-    void Promise.resolve(listener?.(arg)).catch(() => {});
-  } catch {
-    // Synchronous listener throw: contained for the same reason as the rejection above.
-  }
-}
-
 export interface PublishOptions {
   contextGraphId: string;
   quads: Quad[];
